tests(curses): one campaign-mount guard, not three
Deletion, not a behaviour fix: the two copies did the same thing the
shared guard does, so there is no red beyond the tripwire committed
just before this (which named both of them by file and line).

curses_mount_restore.h's MountRestore was a strict SUBSET of
test_curses_picker_client.cpp's ScopedCursesPickerMountRestore — two of
the sixteen picker-client sites call mounted_before() and an idempotent
restore() the shared struct did not have, which is why a rename alone
would not compile. So the shared one is promoted to the superset, taking
the picker-client body verbatim, and both copies go: the picker client's
sixteen declarations become `MountRestore mount_restore;` (the four API
call sites need no change), and curses_network_internal.inc includes the
header instead of carrying a third copy. The .inc is pulled into
src/platform/curses/curses_network.cpp under TESTING, so og_test_curses
and openglad_curses both rebuild.

The header's 18-line rationale block is byte-identical: WP3 moved it
verbatim from test_curses_network.cpp (9dac53d6) and that disposition
stands. The public `before` member the old struct exposed had no readers
(`grep -rn 'mount_restore\.' tests/curses/`), so the nine existing users
are unaffected.

og::test::ScopedCampaignMountState (tests/test_save_state_guard.h:98) is
deliberately left alone: it ADD_FAILUREs on a failed restore, a different
failure surface, and folding it in would change what og_test_curses
reports.

Heritage: the diff deletes ZERO comment lines (the blame audit's step-1
collector finds none in this range), so nothing is owed. Nothing here is
pre-2020 — ScopedCursesPickerMountRestore blames to 430a8022 (2026-07-24)
and the .inc copy to its own 2026 prose.

Green: og_test_curses 314/314 in declaration order (313 before + the
tripwire). Tooth after the consolidation, planted: an uncompiled
tests/curses/planted_twin.h declaring `struct PlantedTwinMountRestore`
puts the tripwire back to red naming planted_twin.h:2.
